tests: Extract report related functions and clean up main file
