+++ tests/linearizability/model/history.go
@@ -16,14 +16,34 @@ package model
 
 import (
 	"fmt"
+	"testing"
 	"time"
 
 	"github.com/anishathalye/porcupine"
+	"go.uber.org/zap"
 
 	"go.etcd.io/etcd/api/v3/etcdserverpb"
 	clientv3 "go.etcd.io/etcd/client/v3"
 	"go.etcd.io/etcd/tests/v3/linearizability/identity"
 )
+
+// ValidateOperationHistoryAndReturnVisualize return visualize as porcupine.linearizationInfo used to generate visualization is private.
+func ValidateOperationHistoryAndReturnVisualize(t *testing.T, lg *zap.Logger, operations []porcupine.Operation) (visualize func(basepath string)) {
+	linearizable, info := porcupine.CheckOperationsVerbose(Etcd, operations, 5*time.Minute)
+	if linearizable == porcupine.Illegal {
+		t.Error("Model is not linearizable")
+	}
+	if linearizable == porcupine.Unknown {
+		t.Error("Linearization timed out")
+	}
+	return func(path string) {
+		lg.Info("Saving visualization", zap.String("path", path))
+		err := porcupine.VisualizePath(Etcd, info, path)
+		if err != nil {
+			t.Errorf("Failed to visualize, err: %v", err)
+		}
+	}
+}
 
 type AppendableHistory struct {
 	// id of the next write operation. If needed a new id might be requested from idProvider.
